Deprecate module "fileformats.ff", with "fileformats.um" as a partial replacement.
