+++ lib/iris/fileformats/um/__init__.py
@@ -1,4 +1,4 @@
-# (C) British Crown Copyright 2014 - 2015, Met Office
+# (C) British Crown Copyright 2014 - 2016, Met Office
 #
 # This file is part of Iris.
 #
@@ -15,9 +15,13 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Iris.  If not, see <http://www.gnu.org/licenses/>.
 """
-Conversion of cubes to/from PP/FF formats.
+Provides iris loading support for UM file types, including FieldsFile.
 
 """
 
 from __future__ import (absolute_import, division, print_function)
 from six.moves import (filter, input, map, range, zip)  # noqa
+
+# Publish the FF-replacement features here, and include documentation.
+from ._ff_replacement import um_to_pp, load_cubes, load_cubes_32bit_ieee
+__all__ = ['um_to_pp', 'load_cubes', 'load_cubes_32bit_ieee']
