+++ lib/iris/tests/unit/fileformats/ff/test_ArakawaC.py
@@ -1,4 +1,4 @@
-# (C) British Crown Copyright 2013 - 2015, Met Office
+# (C) British Crown Copyright 2013 - 2016, Met Office
 #
 # This file is part of Iris.
 #
@@ -25,7 +25,7 @@ import iris.tests as tests
 
 import numpy as np
 
-from iris.fileformats.ff import ArakawaC
+from iris.fileformats._ff import ArakawaC
 
 
 class Test__x_vectors(tests.IrisTest):
